requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
bcrypt>=4.1.0
pyjwt>=2.8.0
qrcode>=7.4.0
pillow>=10.0.0
python-dateutil>=2.8.0
uuid>=1.30
